Refactor payment card classes to use runtime polymorphism and update access modifiers
